Move serializer to form
Update src/Form/Serializer/BaseSerializerHandler.php


Update src/Serializer/SerializerHandlerInterface.php


Update src/Serializer/BaseSerializerHandler.php


Update src/Form/Serializer/BaseSerializerHandler.php


Update src/CoreBundle/Serializer/BaseSerializerHandler.php


Update src/CoreBundle/Serializer/SerializerHandlerInterface.php
