+++ src/CoreBundle/DependencyInjection/SonataCoreExtension.php
@@ -17,13 +17,13 @@ use JMS\Serializer\Handler\SubscribingHandlerInterface;
 use Sonata\CoreBundle\Form\FormHelper;
 use Sonata\CoreBundle\Form\Type\TranslatableChoiceType;
 use Sonata\Doctrine\Bridge\Symfony\Bundle\SonataDoctrineBundle;
+use Sonata\Form\Serializer\BaseSerializerHandler;
 use Sonata\Form\Type\BooleanType;
 use Sonata\Form\Type\CollectionType;
 use Sonata\Form\Type\DateRangeType;
 use Sonata\Form\Type\DateTimeRangeType;
 use Sonata\Form\Type\EqualType;
 use Sonata\Form\Type\ImmutableArrayType;
-use Sonata\Serializer\BaseSerializerHandler;
 use Symfony\Bundle\FrameworkBundle\DependencyInjection\Compiler\FormPass;
 use Symfony\Component\Config\Definition\Processor;
 use Symfony\Component\Config\FileLocator;
